move pkg/nfsinstance to pkg/kcp/nfsinstance
