+++ pkg/kcp/nfsinstance/loadIpRange.go
@@ -3,10 +3,10 @@ package nfsinstance
 import (
 	"context"
 	"fmt"
+	types2 "github.com/kyma-project/cloud-manager/components/kcp/pkg/kcp/nfsinstance/types"
 
 	cloudresourcesv1beta1 "github.com/kyma-project/cloud-manager/components/kcp/api/cloud-control/v1beta1"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/composed"
-	types2 "github.com/kyma-project/cloud-manager/components/kcp/pkg/nfsinstance/types"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
