+++ pkg/kcp/nfsinstance/state.go
@@ -3,7 +3,7 @@ package nfsinstance
 import (
 	cloudresourcesv1beta1 "github.com/kyma-project/cloud-manager/components/kcp/api/cloud-control/v1beta1"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/common/actions/focal"
-	"github.com/kyma-project/cloud-manager/components/kcp/pkg/nfsinstance/types"
+	"github.com/kyma-project/cloud-manager/components/kcp/pkg/kcp/nfsinstance/types"
 )
 
 type state struct {
