+++ pkg/provider/gcp/nfsinstance/state.go
@@ -2,10 +2,10 @@ package nfsinstance
 
 import (
 	"context"
+	"github.com/kyma-project/cloud-manager/components/kcp/pkg/kcp/nfsinstance/types"
 
 	"github.com/kyma-project/cloud-manager/components/kcp/api/cloud-control/v1beta1"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/common/abstractions"
-	"github.com/kyma-project/cloud-manager/components/kcp/pkg/nfsinstance/types"
 	gcpclient "github.com/kyma-project/cloud-manager/components/kcp/pkg/provider/gcp/client"
 	"github.com/kyma-project/cloud-manager/components/kcp/pkg/provider/gcp/nfsinstance/client"
 	"google.golang.org/api/file/v1"
